engine: extract esm/hips_properties.js
