+++ engine/esm/transpiled.js
@@ -3,8 +3,6 @@ import { registerType, registerEnum } from "./typesystem.js";
 
 import {
   Util,
-  Rectangle,
-  Guid,
   Mouse,
 } from "./util.js";
 
@@ -28,7 +26,6 @@ import {
   set_freestandingMode,
   set_globalScriptInterface,
   set_globalWWTControl,
-  set_makeNewHipsProperties,
 } from "./data_globals.js";
 
 import { BlendState } from "./blend_state.js";
@@ -63,7 +60,6 @@ import { Coordinates } from "./coordinates.js";
 import { Annotation, Circle, Poly, PolyLine } from "./annotation.js";
 import { BasePlanets } from "./baseplanets.js";
 import { CameraParameters } from "./camera_parameters.js";
-import { WebFile } from "./web_file.js";
 import { UiTools } from "./ui_tools.js";
 
 import { FitsImage } from "./layers/fits_image.js";
@@ -78,12 +74,10 @@ import { Constellations } from "./constellations.js";
 import { SpaceTimeController } from "./space_time_controller.js";
 import { Planets } from "./planets.js";
 import { Place } from "./place.js";
-import { FolderUp } from "./folder_up.js";
 import { Grids } from "./grids.js";
 import { KeplerVertex } from "./kepler_vertex.js";
 import { MinorPlanets } from "./minor_planets.js";
 import { TileCache } from "./tile_cache.js";
-import { Tour } from "./tour.js";
 import { VideoOutputType } from "./video_output_type.js";
 import { ViewMoverSlew, ViewMoverKenBurnsStyle } from "./view_mover.js";
 import { Dialog } from "./utilities/dialog.js";
@@ -95,7 +89,6 @@ import { ImageSetLayer } from "./layers/imageset_layer.js";
 import { Object3dLayer } from "./layers/object3d.js";
 import { EllipseRenderer } from "./layers/orbit.js";
 import { OrbitLayer } from "./layers/orbit_layer.js";
-import { VoTable } from "./layers/vo_table.js";
 
 import { TourPlayer } from "./tours/tour_player.js";
 
@@ -251,128 +244,6 @@ registerEnum("Formatting", Formatting);
 export function GFX() { }
 
 registerType("GFX", [GFX, null, null]);
-
-// wwtlib.HipsProperties
-
-export function HipsProperties(dataset) {
-  this._properties = {};
-  this._catalogColumnInfo = null;
-  this._catalogSpreadSheetLayer = new CatalogSpreadSheetLayer();
-  this._downloadComplete = false;
-  this.dataset = dataset;
-  this._datasetName = dataset.get_name();
-  this._url = dataset.get_url();
-  if (this._url.toLowerCase().indexOf('norder') > -1) {
-    this._url = this._url.substring(0, this._url.toLowerCase().indexOf('norder'));
-  }
-  this._url += 'properties';
-  this._download();
-}
-
-var HipsProperties$ = {
-  get_properties: function () {
-    return this._properties;
-  },
-
-  get_catalogSpreadSheetLayer: function () {
-    return this._catalogSpreadSheetLayer;
-  },
-
-  set_catalogSpreadSheetLayer: function (value) {
-    this._catalogSpreadSheetLayer = value;
-    return value;
-  },
-
-  get_catalogColumnInfo: function () {
-    return this._catalogColumnInfo;
-  },
-
-  set_catalogColumnInfo: function (value) {
-    this._catalogColumnInfo = value;
-    return value;
-  },
-
-  get_downloadComplete: function () {
-    return this._downloadComplete;
-  },
-
-  _download: function () {
-    this._webFile = new WebFile(this._url);
-    this._webFile.onStateChange = ss.bind('_onPropertiesDownloadComplete', this);
-    this._webFile.send();
-  },
-
-  _onPropertiesDownloadComplete: function () {
-    if (this._webFile.get_state() === 1) {
-      this._parseProperties(this._webFile.getText());
-      if (ss.keyExists(this.get_properties(), 'dataproduct_type') && this.get_properties()['dataproduct_type'].toLowerCase() === 'catalog') {
-        this._catalogColumnInfo = VoTable.loadFromUrl(ss.replaceString(this._url, '/properties', '/metadata.xml'), ss.bind('_onCatalogMetadataDownloadComplete', this));
-      }
-      else {
-        if (ss.keyExists(this.get_properties(), 'hips_data_range')) {
-          var hips_data_range = this.get_properties()['hips_data_range'];
-          this.dataset.get_fitsProperties().minVal = parseFloat(hips_data_range.split(' ')[0]);
-          this.dataset.get_fitsProperties().maxVal = parseFloat(hips_data_range.split(' ')[1]);
-          this.dataset.get_fitsProperties().lowerCut = this.dataset.get_fitsProperties().minVal;
-          this.dataset.get_fitsProperties().upperCut = this.dataset.get_fitsProperties().maxVal;
-        }
-        if (ss.keyExists(this.get_properties(), 'hips_pixel_cut')) {
-          var hips_pixel_cut = this.get_properties()['hips_pixel_cut'];
-          this.dataset.get_fitsProperties().lowerCut = parseFloat(hips_pixel_cut.split(' ')[0]);
-          this.dataset.get_fitsProperties().upperCut = parseFloat(hips_pixel_cut.split(' ')[1]);
-          if (!ss.keyExists(this.get_properties(), 'hips_data_range')) {
-            this.dataset.get_fitsProperties().minVal = this.dataset.get_fitsProperties().lowerCut;
-            this.dataset.get_fitsProperties().maxVal = this.dataset.get_fitsProperties().upperCut;
-          }
-        }
-        this._downloadComplete = true;
-        if (this._onDownloadComplete != null) {
-          this._onDownloadComplete();
-        }
-      }
-    }
-  },
-
-  _onCatalogMetadataDownloadComplete: function () {
-    this._catalogSpreadSheetLayer.useHeadersFromVoTable(this._catalogColumnInfo);
-    this._catalogSpreadSheetLayer.set_name(this._datasetName);
-    this._catalogSpreadSheetLayer.id = Guid.createFrom(this._datasetName);
-    LayerManager.addSpreadsheetLayer(this.get_catalogSpreadSheetLayer(), 'Sky');
-    this._downloadComplete = true;
-    if (this._onDownloadComplete != null) {
-      this._onDownloadComplete();
-    }
-  },
-
-  setDownloadCompleteListener: function (listener) {
-    this._onDownloadComplete = listener;
-  },
-
-  _parseProperties: function (data) {
-    var lines = data.split('\n');
-    var $enum1 = ss.enumerate(lines);
-    while ($enum1.moveNext()) {
-      var line = $enum1.current;
-      if (!ss.whitespace(line) && !ss.startsWith(line, '#')) {
-        var parts = line.split('=');
-        if (parts.length === 2) {
-          var key = ss.trim(parts[0]);
-          var val = ss.trim(parts[1]);
-          if (!ss.whitespace(key) && !ss.whitespace(val)) {
-            this.get_properties()[key] = val;
-          }
-        }
-      }
-    }
-  }
-};
-
-registerType("HipsProperties", [HipsProperties, HipsProperties$, null]);
-
-set_makeNewHipsProperties(function (imageset) {
-  return new HipsProperties(imageset);
-});
-
 
 // The `Layer.fromXml` function, which needs to be aware of a bunch of layer
 // subclasses.
